Create class spi connector to improve separation of concern
